% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_bessel.R
\name{Bessel}
\alias{Bessel}
\alias{dbessel}
\alias{pbessel}
\alias{qbessel}
\alias{rbessel}
\title{Bessel Distribution}
\usage{
dbessel(x, mu, phi, log = FALSE)

pbessel(q, mu, phi, lower.tail = TRUE, log.p = FALSE)

qbessel(p, mu, phi, lower.tail = TRUE, log.p = FALSE)

rbessel(n, mu, phi)
}
\arguments{
\item{x, q}{vector of quantiles (0 < x, q < 1).}

\item{mu}{mean parameter (0 < mu < 1).}

\item{phi}{precision parameter (phi > 0).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\description{
Density, distribution function, quantile function and random generation for 
the Bessel regression distribution with mean \code{mu} and precision \code{phi}.
}
\details{
The probability density function for the Bessel distribution is given by:
\deqn{f(y \mid \mu, \phi) = \frac{\mu(1-\mu) \phi e^{\phi}}{\pi [y(1-y)]^{3/2} \zeta_{\mu}(y)} K_1(\phi \zeta_{\mu}(y))}
where \eqn{\zeta_{\mu}(y) = \sqrt{1 + \frac{(y-\mu)^2}{y(1-y)}}} and \eqn{K_1} is the modified Bessel function of the second kind.

The random generation utilizes a highly efficient C++ implementation of the Inverse Gaussian mixture property proposed by Giner and Smyth (2016).
}
\examples{
# 1. Density Evaluation
dbessel(0.5, mu = 0.5, phi = 10)

# 2. Cumulative Probability
pbessel(0.5, mu = 0.5, phi = 10)

# 3. Random Generation & Histogram
set.seed(123)
r <- rbessel(1000, mu = 0.6, phi = 5)
hist(r, prob = TRUE, breaks = 30, col = "grey", main = "Bessel(0.6, 5)")
curve(dbessel(x, mu = 0.6, phi = 5), add = TRUE, col = "red", lwd = 2)

}
\references{
Barreto-Souza, W., Mayrink, V. D., & Simas, A. B. (2022). 
Bessel regression model: a robust alternative to beta regression. 
\emph{Journal of Applied Statistics}, \bold{49}(1), 1--20.
}
\author{
Ricardo Rasmussen Petterle
}
