% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.unitregTMB}
\alias{coef.unitregTMB}
\title{Extract Model Coefficients}
\usage{
\method{coef}{unitregTMB}(object, type = c("all", "mu", "phi", "p0", "p1"), ...)
}
\arguments{
\item{object}{An object of class \code{unitregTMB}.}

\item{type}{A character string indicating the type of coefficients to extract. 
Available options are \code{"all"} (default), \code{"mu"} (location), 
\code{"phi"} (precision/dispersion), \code{"p0"} (zero-inflation), or \code{"p1"} (one-inflation).}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A numeric vector of estimated coefficients.
}
\description{
Extracts the estimated coefficients from a fitted \code{unitregTMB} model.
}
