% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{df.residual.unitregTMB}
\alias{df.residual.unitregTMB}
\title{Extract the Residual Degrees of Freedom}
\usage{
\method{df.residual}{unitregTMB}(object, ...)
}
\arguments{
\item{object}{An object of class \code{unitregTMB}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An integer representing the residual degrees of freedom.
}
\description{
Extracts the residual degrees of freedom from a fitted \code{unitregTMB} model.
}
