% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fixed_effects}
\alias{fixed_effects}
\alias{fixed_effects.unitregTMB}
\title{Extract Fixed Effects}
\usage{
fixed_effects(object, ...)

\method{fixed_effects}{unitregTMB}(object, ...)
}
\arguments{
\item{object}{A fitted object of class "unitregTMB".}

\item{...}{Additional arguments.}
}
\value{
A numeric vector of fixed effects.
}
\description{
Generic function to extract fixed effects from fitted models.
}
