% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fixef.unitregTMB}
\alias{fixef.unitregTMB}
\title{Extract Fixed-Effects Estimates}
\usage{
fixef.unitregTMB(object, ...)
}
\arguments{
\item{object}{An object of class \code{unitregTMB}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class \code{fixef.unitregTMB} containing the estimated fixed-effects.
}
\description{
Extracts the fixed-effects estimates from a fitted \code{unitregTMB} model.
}
