% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gof}
\alias{gof}
\title{Goodness-of-Fit Measures}
\usage{
gof(object, ..., digits = 2)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional fitted model objects.}

\item{digits}{Number of significant digits to be used in the output.}
}
\description{
Generic function to compute and compile goodness-of-fit measures for regression models.
}
