% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gof.unitregTMB}
\alias{gof.unitregTMB}
\title{Goodness-of-Fit for unitregTMB Models}
\usage{
\method{gof}{unitregTMB}(object, ..., digits = 2)
}
\arguments{
\item{object}{An object of class \code{unitregTMB}.}

\item{...}{Additional \code{unitregTMB} model objects for comparison.}

\item{digits}{Number of significant digits to be used in the output.}
}
\value{
An object of class \code{gof.unitregTMB} containing a list of computed metrics for each model.
}
\description{
Computes and compares log-likelihood, AIC, and BIC for one or more fitted \code{unitregTMB} models.
}
