% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gof_tab}
\alias{gof_tab}
\alias{gof_tab.unitregTMB}
\title{Goodness-of-Fit Measures}
\usage{
gof_tab(object, ..., digits = 2)

\method{gof_tab}{unitregTMB}(object, ..., digits = 2)
}
\arguments{
\item{object}{An object of class \code{unitregTMB}.}

\item{...}{Additional \code{unitregTMB} model objects for comparison.}

\item{digits}{Number of significant digits to be used in the output.}
}
\value{
An object of class \code{gof_tab.unitregTMB} containing a list of computed metrics for each model.
}
\description{
Generic function to compute and compile goodness-of-fit measures for regression models.

Computes and compares log-likelihood, AIC, and BIC for one or more fitted \code{unitregTMB} models.
}
\examples{
\donttest{
# Assuming fit1 and fit2 are two fitted unitregTMB models:
# fit1 <- unitregTMB(Y ~ educ, data = da, family = vasicek())
# fit2 <- unitregTMB(Y ~ educ, data = da, family = kumaraswamy())
# gof_tab(fit1, fit2)
}
}
