% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gof_tab.unitregTMB}
\alias{gof_tab.unitregTMB}
\alias{print.gof_tab.unitregTMB}
\title{Goodness-of-Fit Table for unitregTMB Models}
\usage{
\method{gof_tab}{unitregTMB}(object, ..., digits = 2)

\method{print}{gof_tab.unitregTMB}(x, digits = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{unitregTMB}.}

\item{...}{Additional \code{unitregTMB} model objects for comparison.}

\item{digits}{Number of significant digits to be used in the output.}

\item{x}{An object of class \code{gof_tab.unitregTMB}.}
}
\value{
An object of class \code{gof_tab.unitregTMB} containing a list of computed metrics for each model.
}
\description{
Computes and compares log-likelihood, AIC, and BIC for one or more fitted \code{unitregTMB} models.
}
\examples{
\donttest{
# Assuming fit1 and fit2 are two fitted unitregTMB models:
# fit1 <- unitregTMB(arms ~ age + bmi + gender + ipaq, 
                     family = vasicek(model_for = "quantile", tau = 0.75), 
                     data = bodyfat)
# fit2 <- unitregTMB(arms ~ age + bmi + gender + ipaq, 
                     family = kumaraswamy(model_for = "quantile", tau = 0.75), 
                     data = bodyfat)
# gof_tab(fit1, fit2)
}
}
