% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnp.R
\name{hnp.unitregTMB}
\alias{hnp.unitregTMB}
\title{Half-Normal Plots with Simulated Envelopes for unitregTMB}
\usage{
\method{hnp}{unitregTMB}(
  object,
  nsim = 99,
  halfnormal = TRUE,
  plot = TRUE,
  level = 0.95,
  resid.type = c("quantile", "cox-snell"),
  ncpus = 1,
  ...
)
}
\arguments{
\item{object}{A fitted \code{unitregTMB} model.}

\item{nsim}{Number of simulations used to compute the envelope. Default is 99.}

\item{halfnormal}{logical. If \code{TRUE} (default), a half-normal plot is produced. 
If \code{FALSE}, a normal plot is produced.}

\item{plot}{logical. Should the plot be printed? Default is \code{TRUE}.}

\item{level}{Confidence level of the simulated envelope. Default is 0.95.}

\item{resid.type}{Type of residuals to be used: \code{"quantile"} (default) or \code{"cox-snell"}.}

\item{ncpus}{Number of cores to use for parallel computing. Default is 1.}

\item{...}{Additional arguments passed to the \code{plot} function.}
}
\value{
A list (invisibly) containing the observed residuals (\code{res_obs}) and a matrix 
        with the lower and upper bounds of the simulated envelope (\code{envelope}).
}
\description{
Produces a (half-)normal probability plot with a simulated envelope for a fitted
\code{unitregTMB} model. This diagnostic tool assesses the goodness-of-fit of the 
chosen continuous distribution as well as the zero-one inflation (ZOI) components.
}
