% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs.unitregTMB}
\alias{nobs.unitregTMB}
\title{Extract the Number of Observations}
\usage{
\method{nobs}{unitregTMB}(object, ...)
}
\arguments{
\item{object}{An object of class \code{unitregTMB}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An integer representing the number of observations.
}
\description{
Extracts the number of observations used to fit a \code{unitregTMB} model.
}
