% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuong_test.R
\name{pairwise_vuong_test.unitregTMB}
\alias{pairwise_vuong_test.unitregTMB}
\title{Pairwise Vuong Test for Multiple unitregTMB Models}
\usage{
\method{pairwise_vuong_test}{unitregTMB}(
  ...,
  lt = NULL,
  p.adjust.method = stats::p.adjust.methods,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{...}{\code{unitregTMB} objects separated by commas.}

\item{lt}{A list containing one or more \code{unitregTMB} objects.}

\item{p.adjust.method}{A character string specifying the method for multiple
testing adjustment; almost always one of \code{p.adjust.methods}. Can be abbreviated.}

\item{alternative}{Indicates the alternative hypothesis and must be one
of \code{"two.sided"} (default), \code{"less"}, or \code{"greater"}. Can be abbreviated.}
}
\value{
An object of class \code{"pairwise_htest"}
}
\description{
Calculates pairwise comparisons between multiple fitted models, 
performing the Vuong test for objects of class \code{unitregTMB}.
}
\examples{
\donttest{
sim_data <- sim_unitreg(n_id = 500, family = vasicek(), formula_mu = ~ x, 
                        beta_mu = c(0.5, -0.5), beta_phi = 1.5,
                        cov_config = list(x = list(dist = "normal")))

fit_vas <- unitregTMB(Y ~ x, data = sim_data, family = vasicek())
fit_bet <- unitregTMB(Y ~ x, data = sim_data, family = beta_fam())
fit_kum <- unitregTMB(Y ~ x, data = sim_data, family = kumaraswamy())

ans <- pairwise_vuong_test(fit_vas, fit_bet, fit_kum)
print(ans)
}

}
\seealso{
\code{\link{vuong_test}}, \code{\link[stats]{p.adjust}}
}
