% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{periodontal}
\alias{periodontal}
\title{Periodontal Disease Data Set}
\format{
A data frame with 1160 rows and 8 variables:
\describe{
  \item{id}{Subject identifier.}
  \item{tooth}{Tooth types (molar; premolar; canine; incisor).}
  \item{y}{The response variable lies on the closed interval [0,1] and represents the proportion of diseased tooth sites for each of the four tooth types. Contains exact zeros and ones.}
  \item{gender}{Patient gender (Male; Female).}
  \item{age}{Patient age in years.}
  \item{hba1c}{Glycosylated hemoglobin status indicator (Controlled, <7\%; Uncontrolled, >=7\%).}
  \item{smoker}{Smoking status (Non-smoker; Smoker).}
  \item{bmi}{Body mass index.}
}
}
\source{
Downloaded from Dr. Bandyopadhyay's software page: \url{https://sites.google.com/view/dbandyop/software}.
Original dataset archive: \url{https://drive.google.com/file/d/1-fyruHi5eMp6TvFegQazx5lios6igLGc/view?pli=1}.
}
\usage{
periodontal
}
\description{
The data come from a study conducted at the Medical University of South Carolina by Fernandes et al. (2006).
The data set involves 1160 clustered observations from 290 subjects.
}
\examples{
data("periodontal", package = "unitregTMB")
head(periodontal)
\donttest{
# Example: Fitting a zero-and-one augmented mixed-effects model
# fit <- unitregTMB(y ~ age + gender + tooth + (1 | id), 
#                   data = periodontal, 
#                   family = unitgamma())
}
}
\references{
Zhao, W., Lian, H., & Bandyopadhyay, D. (2018). A partially linear additive model for clustered proportion data. \emph{Statistics in Medicine}, 37(6), 1009-1030.

Fernandes, J., et al. (2006). Prevalence of periodontal disease in Gullah African American diabetics. \emph{Journal of Dental Research}, 85(Special Issue A), 997.
}
\keyword{datasets}
