% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{print.anova.unitregTMB}
\alias{print.anova.unitregTMB}
\title{Print Method for anova.unitregTMB Objects}
\usage{
\method{print}{anova.unitregTMB}(x, ...)
}
\arguments{
\item{x}{An object of class \code{anova.unitregTMB}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the original \code{anova.unitregTMB} object.
}
\description{
Custom print method for displaying the analysis of deviance table 
generated by \code{anova.unitregTMB}.
}
