% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.fixef.unitregTMB}
\alias{print.fixef.unitregTMB}
\title{Print Method for fixef.unitregTMB}
\usage{
\method{print}{fixef.unitregTMB}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class \code{fixef.unitregTMB}.}

\item{digits}{Number of significant digits to use when printing.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the original \code{fixef.unitregTMB} object.
}
\description{
Custom print method for \code{fixef.unitregTMB} objects.
}
