% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.gof.unitregTMB}
\alias{print.gof.unitregTMB}
\title{Print Goodness-of-Fit Measures}
\usage{
\method{print}{gof.unitregTMB}(x, digits = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{gof.unitregTMB}.}

\item{digits}{Number of significant digits to use when printing.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the original \code{gof.unitregTMB} object.
}
\description{
Prints the compiled goodness-of-fit table for \code{unitregTMB} models.
}
