% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.unitregTMB}
\alias{print.summary.unitregTMB}
\title{Print Summary for unitregTMB Models}
\usage{
\method{print}{summary.unitregTMB}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class \code{summary.unitregTMB}.}

\item{digits}{The number of significant digits to use when printing.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the original \code{summary.unitregTMB} object.
}
\description{
Custom print method for \code{summary.unitregTMB} objects.
}
