% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_coef.R
\name{print.summary_coef.unitregTMB}
\alias{print.summary_coef.unitregTMB}
\title{Print Method for summary_coef.unitregTMB Objects}
\usage{
\method{print}{summary_coef.unitregTMB}(x, digits = NULL, print_tex = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{summary_coef.unitregTMB}.}

\item{digits}{Number of decimal places to use when printing. If \code{NULL}, 
it inherits the value specified during the \code{summary_coef} call.}

\item{print_tex}{A logical indicating whether to print the table in LaTeX format. 
If \code{NULL}, it inherits the value specified during the \code{summary_coef} call.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the original \code{summary_coef.unitregTMB} object.
}
\description{
Prints the compiled coefficient table generated by \code{summary_coef.unitregTMB}. 
The output can be displayed as a standard console table or generated as a LaTeX 
formatted table for easy inclusion in academic manuscripts.
}
