% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effects.R
\name{random_effects.unitregTMB}
\alias{random_effects.unitregTMB}
\title{Extract Random Effects (BLUPs) from a unitregTMB Object}
\usage{
\method{random_effects}{unitregTMB}(object, ...)
}
\arguments{
\item{object}{A fitted object of class "unitregTMB".}

\item{...}{Additional arguments (not used).}
}
\value{
A list (one element per random term/group) 
        containing data.frames of the random effects.
}
\description{
Extracts the Best Linear Unbiased Predictors (BLUPs) for the random 
effects of a fitted \code{unitregTMB} model.
}
\examples{
\donttest{
# Assuming 'da' is your dataset and 'id' is the grouping variable:
# fit <- unitregTMB(Y ~ educ + (1 | id), data = da, family = vasicek())
# 
# # Extract the random effects as a list of data.frames
# re <- random_effects(fit)
# head(re$id)
}
}
