% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{ranef.unitregTMB}
\alias{ranef.unitregTMB}
\title{Extract Random Effects (BLUPs) from a unitregTMB Object}
\usage{
ranef.unitregTMB(object, ...)
}
\arguments{
\item{object}{A fitted object of class "unitregTMB".}

\item{...}{Additional arguments (not used).}
}
\value{
A list (one element per random term/group) 
        containing data.frames of the random effects.
}
\description{
Extract Random Effects (BLUPs) from a unitregTMB Object
}
