% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.unitregTMB}
\alias{simulate.unitregTMB}
\title{Simulate Responses from a Fitted unitregTMB Model}
\usage{
\method{simulate}{unitregTMB}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{unitregTMB} object.}

\item{nsim}{Number of response vectors to simulate. Defaults to 1.}

\item{seed}{An object specifying if and how the random number generator should be initialized.}

\item{...}{Additional optional arguments.}
}
\value{
A data.frame with \code{nsim} columns containing the simulated responses.
}
\description{
Generates simulated response data from a fitted \code{unitregTMB} model. 
This function perfectly mimics the data-generating process (DGP) estimated by the model, 
including the continuous distribution and any zero/one inflation parameters.
}
