% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_coef.R
\name{summary_coef}
\alias{summary_coef}
\title{Extract and Summarize Model Coefficients}
\usage{
summary_coef(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic function to extract, format, and compile coefficients and standard errors 
from fitted model objects.
}
