% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_unitgompertz.R
\name{UnitGompertz}
\alias{UnitGompertz}
\alias{dugompertz_mode}
\alias{pugompertz_mode}
\alias{qugompertz_mode}
\alias{rugompertz_mode}
\alias{dugompertz_quantile}
\alias{pugompertz_quantile}
\alias{qugompertz_quantile}
\alias{rugompertz_quantile}
\title{Unit-Gompertz Distribution}
\usage{
dugompertz_mode(x, mu, phi, log = FALSE)

pugompertz_mode(q, mu, phi, lower.tail = TRUE, log.p = FALSE)

qugompertz_mode(p, mu, phi, lower.tail = TRUE, log.p = FALSE)

rugompertz_mode(n, mu, phi)

dugompertz_quantile(x, mu, phi, tau = 0.5, log = FALSE)

pugompertz_quantile(q, mu, phi, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

qugompertz_quantile(p, mu, phi, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

rugompertz_quantile(n, mu, phi, tau = 0.5)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{mode or quantile parameter, depending on the function used (0 < mu < 1).}

\item{phi}{shape parameter (phi > 0).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{tau}{the quantile level (0 < tau < 1). Required only for quantile parameterizations.}
}
\description{
Density, distribution function, quantile function, and random number generation for the 
Unit-Gompertz distribution parameterized either by the Mode or a given Quantile.
}
\examples{
# 1. Mode Parameterization
mu_mode <- 0.5
phi_val <- 2.0
x1 <- rugompertz_mode(1000, mu_mode, phi_val)
hist(x1, prob = TRUE, main = "Unit-Gompertz (Mode)", col = "lightgray")
curve(dugompertz_mode(x, mu_mode, phi_val), add = TRUE, col = "red", lwd = 2)

# 2. Quantile Parameterization (Median)
mu_quant <- 0.7
phi_val <- 1.5
x2 <- rugompertz_quantile(1000, mu_quant, phi_val, tau = 0.5)
hist(x2, prob = TRUE, main = "Unit-Gompertz (Median)", col = "lightblue")
curve(dugompertz_quantile(x, mu_quant, phi_val, tau = 0.5), add = TRUE, col = "blue", lwd = 2)

}
\references{
Mazucheli, J., Menezes, A. F. B., & Dey, S. (2019). The unit-Gompertz distribution with applications. 
\emph{Statistica}, \bold{79}(1), 25--43.

Menezes, A. F., Mazucheli, J. and Chakraborty, S. (2021). A collection of parametric modal regression models for bounded data.
\emph{Journal of Biopharmaceutical Statistics}, \bold{31}(4), 490--506.
}
\author{
Ricardo Rasmussen Petterle
}
