% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{unitregTMB-families}
\alias{unitregTMB-families}
\alias{beta_fam}
\alias{simplex}
\alias{vasicek}
\alias{unitgamma}
\alias{bessel}
\alias{kumaraswamy}
\alias{unitgompertz}
\alias{unitweibull}
\alias{johnsonsb}
\alias{ashw}
\alias{ubs}
\title{Family Objects for unitregTMB Models}
\usage{
beta_fam(model_for = c("mean", "mode"), link = "logit")

simplex(model_for = c("mean"), link = "logit")

vasicek(model_for = c("mean", "quantile"), link = "logit", tau = 0.5)

unitgamma(model_for = c("mean", "mode"), link = "logit")

bessel(model_for = c("mean"), link = "logit")

kumaraswamy(model_for = c("mode", "quantile"), link = "logit", tau = 0.5)

unitgompertz(model_for = c("mode", "quantile"), link = "logit", tau = 0.5)

unitweibull(model_for = c("quantile"), tau = 0.5, link = "logit")

johnsonsb(model_for = c("quantile"), tau = 0.5, link = "logit")

ashw(model_for = c("quantile"), tau = 0.5, link = "logit")

ubs(model_for = c("quantile"), tau = 0.5, link = "logit")
}
\description{
Functions to specify the distribution family and link function for \code{unitregTMB} models.
}
