% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitregTMB.R
\name{unitregTMB.control}
\alias{unitregTMB.control}
\title{Control Parameters for unitregTMB Model Fitting}
\usage{
unitregTMB.control(
  eval.max = 2000,
  iter.max = 2000,
  verbose = FALSE,
  sd.report = TRUE,
  get.joint.precision = FALSE,
  optimizer = c("nlminb", "optim"),
  optim.method = "BFGS",
  optim.control = list(),
  ...
)
}
\arguments{
\item{eval.max}{Maximum number of evaluations of the objective function allowed. Default is 2000.}

\item{iter.max}{Maximum number of iterations allowed. Default is 2000.}

\item{verbose}{Logical; if \code{TRUE}, optimization tracing information is printed. Default is \code{FALSE}.}

\item{sd.report}{Logical; if \code{TRUE}, standard errors are calculated using \code{TMB::sdreport}. Default is \code{TRUE}.}

\item{get.joint.precision}{Logical; if \code{TRUE}, the joint precision matrix is returned by \code{sdreport}. Default is \code{FALSE}.}

\item{optimizer}{A character string specifying the optimizer to use. Options are \code{"nlminb"} (default) or \code{"optim"}.}

\item{optim.method}{The method to be used if \code{optimizer = "optim"}. Default is \code{"BFGS"}.}

\item{optim.control}{A list of additional control parameters to be passed to the chosen optimizer.}

\item{...}{Additional arguments.}
}
\value{
A list containing the specified control parameters to be used in \code{unitregTMB.fit}.
}
\description{
Defines control parameters for the optimization process used in \code{unitregTMB}. 
It allows the user to specify evaluation and iteration limits, the choice of 
optimizer, and whether to compute standard errors via the TMB framework.
}
