% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_unitweibull.R
\name{UnitWeibull}
\alias{UnitWeibull}
\alias{duweibull}
\alias{puweibull}
\alias{quweibull}
\alias{ruweibull}
\title{The Unit-Weibull Distribution}
\usage{
duweibull(x, mu, phi, tau = 0.5, log = FALSE)

puweibull(q, mu, phi, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

quweibull(p, mu, phi, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

ruweibull(n, mu, phi, tau = 0.5)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{quantile parameter (0 < mu < 1).}

\item{phi}{shape parameter (phi > 0).}

\item{tau}{the quantile level (0 < tau < 1). Defaults to 0.5 (median).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function, and random number generation for the 
Unit-Weibull distribution parameterized by a quantile \code{mu} at level \code{tau} and shape \code{phi}.
}
\examples{
# 1. Set parameters (Median parameterization)
mu_med <- 0.6
phi_val <- 2.5

# 2. Random generation
set.seed(123)
x <- ruweibull(1000, mu_med, phi_val, tau = 0.5)

# 3. Visualization
hist(x, prob = TRUE, main = "Unit-Weibull (Median)", col = "lightgray")
curve(duweibull(x, mu_med, phi_val, tau = 0.5), add = TRUE, col = "red", lwd = 2)

# 4. Check P(X <= mu) = tau
p_mu <- puweibull(mu_med, mu_med, phi_val, tau = 0.5)
cat("P(X <= mu) should be 0.5:", p_mu, "\n")

}
\references{
Mazucheli, J., Menezes, A. F. B., & Ghitany, M. E. (2018). The unit-Weibull distribution and associated inference. 
\emph{Journal of Applied Statistics}, \bold{45}(8), 1351--1368.
}
\author{
Ricardo Rasmussen Petterle
}
