% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_vasicek.R
\name{Vasicek}
\alias{Vasicek}
\alias{dvasicek_mean}
\alias{pvasicek_mean}
\alias{qvasicek_mean}
\alias{rvasicek_mean}
\alias{dvasicek_quantile}
\alias{pvasicek_quantile}
\alias{qvasicek_quantile}
\alias{rvasicek_quantile}
\title{Vasicek Distribution}
\usage{
dvasicek_mean(x, mu, phi, log = FALSE)

pvasicek_mean(q, mu, phi, lower.tail = TRUE, log.p = FALSE)

qvasicek_mean(p, mu, phi, lower.tail = TRUE, log.p = FALSE)

rvasicek_mean(n, mu, phi)

dvasicek_quantile(x, mu, phi, tau = 0.5, log = FALSE)

pvasicek_quantile(q, mu, phi, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

qvasicek_quantile(p, mu, phi, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

rvasicek_quantile(n, mu, phi, tau = 0.5)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{mean or quantile parameter, depending on the function used (0 < mu < 1).}

\item{phi}{shape parameter (0 < phi < 1). See Details.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{tau}{the quantile level (0 < tau < 1). Required only for quantile parameterizations.}
}
\description{
Density, distribution function, quantile function, and random number generation for the 
Vasicek distribution parameterized either by the Mean or a given Quantile.
}
\details{
The Vasicek distribution is bounded on the unit interval. In this implementation, the 
shape parameter \code{phi} corresponds directly to the standard \eqn{\theta \in (0,1)} parameter 
from the literature. The mapping from the unconstrained real line is handled upstream 
by the \code{unitregTMB} model via a logit link.
}
\examples{
# 1. Mean Parameterization
mu_m <- 0.5
phi_m <- 0.2 # Corresponds directly to theta = 0.2
x1 <- rvasicek_mean(1000, mu_m, phi_m)
hist(x1, prob = TRUE, main = "Vasicek (Mean)", col = "lightgray", border = "white")
curve(dvasicek_mean(x, mu_m, phi_m), add = TRUE, col = "red", lwd = 2)

# 2. Quantile Parameterization (Median)
mu_q <- 0.7
phi_q <- 0.5 # Corresponds directly to theta = 0.5
x2 <- rvasicek_quantile(1000, mu_q, phi_q, tau = 0.5)
hist(x2, prob = TRUE, main = "Vasicek (Median)", col = "lightblue", border = "white")
curve(dvasicek_quantile(x, mu_q, phi_q, tau = 0.5), add = TRUE, col = "blue", lwd = 2)

}
\references{
Mazucheli, J., Alves, B., Korkmaz, M. C., & Leiva, V. (2022). Vasicek Quantile and Mean 
Regression Models for Bounded Data: New Formulation, Mathematical Derivations, and 
Numerical Applications. \emph{Mathematics}, \bold{10}(9), 1389.
}
\author{
Ricardo Rasmussen Petterle
}
