% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuong_test.R
\name{vuong_test.unitregTMB}
\alias{vuong_test.unitregTMB}
\title{Vuong Test for Non-Nested unitregTMB Models}
\usage{
\method{vuong_test}{unitregTMB}(
  object1,
  object2,
  alternative = c("two.sided", "less", "greater"),
  ...
)
}
\arguments{
\item{object1, object2}{Objects of class \code{unitregTMB} containing the fitted models.}

\item{alternative}{Indicates the alternative hypothesis and must be one
of \code{"two.sided"} (default), \code{"less"}, or \code{"greater"}. You can
specify just the initial letter of the value.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A list with class \code{"htest"} containing the following components:
\item{statistic}{the value of the test statistic.}
\item{p.value}{the p-value of the test.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string with the method used.}
\item{data.name}{a character string given the name of families models under comparison.}
}
\description{
Performs the Vuong likelihood ratio test for model selection 
between two fitted non-nested models of class \code{unitregTMB}.
}
\details{
The statistic of Vuong likelihood ratio test for compare two
non-nested regression models is defined by
\deqn{T = \frac{1}{\widehat{\omega}^2\,\sqrt{n}}\,\sum_{i=1}^{n}\,
\log\frac{f(y_i \mid \boldsymbol{x}_i, \widehat{\boldsymbol{\theta}})}{
g(y_i \mid \boldsymbol{x}_i,\widehat{\boldsymbol{\gamma}})}}
where
\deqn{\widehat{\omega}^2 = \frac{1}{n}\,\sum_{i=1}^{n}\,\left(\log \frac{f(y_i \mid \boldsymbol{x}_i, \widehat{\boldsymbol{\theta}})}{g(y_i \mid \boldsymbol{x}_i, \widehat{\boldsymbol{\gamma}})}\right)^2 - \left[\frac{1}{n}\,\sum_{i=1}^{n}\,\left(\log \frac{f(y_i \mid \boldsymbol{x}_i, \widehat{\boldsymbol{\theta}})}{ g(y_i \mid \boldsymbol{x}_i, \widehat{\boldsymbol{\gamma}})}\right)\right]^2}
is an estimator for the variance of the log-likelihood ratio.

When \eqn{n \rightarrow \infty} we have that \eqn{T \rightarrow N(0, 1)} in distribution.
Therefore, at \eqn{\alpha\%} level of significance, the null hypothesis of
the equivalence of the competing models is rejected if \eqn{|T| > z_{\alpha/2}},
where \eqn{z_{\alpha/2}} is the \eqn{\alpha/2} quantile of standard normal distribution.

In practical terms, \eqn{f(y_i \mid \boldsymbol{x}_i, \widehat{\boldsymbol{\theta}})}
is better (worse) than \eqn{g(y_i \mid \boldsymbol{x}_i, \widehat{\boldsymbol{\gamma}})}
if \eqn{T > z_{\alpha/2}} (or \eqn{T < -z_{\alpha/2}}).
}
\examples{
\donttest{
# Simulate independent data
sim_data <- sim_unitreg(n_id = 500, family = vasicek(), formula_mu = ~ x, 
                        beta_mu = c(0.5, -0.5), beta_phi = 1.5,
                        cov_config = list(x = list(dist = "normal")))

# Fit two competing models
fit_vasicek <- unitregTMB(Y ~ x, data = sim_data, family = vasicek())
fit_beta <- unitregTMB(Y ~ x, data = sim_data, family = beta_fam())

# Perform Vuong test
ans <- vuong_test(fit_vasicek, fit_beta)
print(ans)
}

}
\references{
Vuong, Q. (1989). Likelihood ratio tests for model selection and
non-nested hypotheses. \emph{Econometrica}, \bold{57}(2), 307--333.
}
\author{
Ricardo Rasmussen Petterle
}
