% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_methods.R
\name{write_methods}
\alias{write_methods}
\title{Generate Methods Section Text for unitregTMB}
\usage{
write_methods(model, language = c("en", "pt"))
}
\description{
Generate Methods Section Text for unitregTMB
}
