% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{AIC.unitregTMB}
\alias{AIC.unitregTMB}
\title{Extract Akaike Information Criterion}
\usage{
\method{AIC}{unitregTMB}(object, ..., k = 2)
}
\arguments{
\item{object}{An object of class \code{unitregTMB}.}

\item{...}{Further arguments passed to or from other methods.}

\item{k}{Numeric, the penalty per parameter to be used; the default \code{k = 2} 
is the classical AIC.}
}
\value{
A numeric value representing the AIC.
}
\description{
Extracts the Akaike Information Criterion (AIC) of a fitted \code{unitregTMB} model.
}
