% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.unitregTMB}
\alias{anova.unitregTMB}
\title{Likelihood Ratio Tests for unitregTMB Models}
\usage{
\method{anova}{unitregTMB}(object, ..., test = "Chisq")
}
\arguments{
\item{object}{A fitted model object of class \code{unitregTMB}.}

\item{...}{Additional fitted model objects of class \code{unitregTMB} to be compared. 
At least one additional model must be provided.}

\item{test}{A character string specifying the test statistic to use. Currently, 
the only supported option is \code{"Chisq"}, which computes the Likelihood 
Ratio Test based on the Chi-square distribution.}
}
\value{
An object of class \code{anova.unitregTMB} inheriting from \code{data.frame}. 
  The returned data frame contains the following columns:
  \item{Df}{The number of estimated parameters in the model.}
  \item{AIC}{Akaike Information Criterion.}
  \item{BIC}{Bayesian Information Criterion.}
  \item{logLik}{The log-likelihood of the model.}
  \item{deviance}{The deviance of the model (-2 * log-likelihood).}
  \item{Chisq}{The test statistic for the likelihood ratio test.}
  \item{Chi Df}{The difference in degrees of freedom between the models.}
  \item{Pr(>Chisq)}{The p-value associated with the likelihood ratio test.}
}
\description{
Computes an analysis of deviance table for two or more fitted \code{unitregTMB} 
models. This function performs a Likelihood Ratio Test (LRT) to compare nested models, 
providing statistics such as degrees of freedom, AIC, BIC, log-likelihood, 
and the corresponding p-value.
}
\examples{
\dontrun{
# Assuming fit1 and fit2 are fitted unitregTMB models
# fit1 <- unitregTMB(y ~ x1, data = da, family = "beta")
# fit2 <- unitregTMB(y ~ x1 + x2, data = da, family = "beta")
# anova(fit1, fit2)
}

}
\author{
Ricardo Rasmussen Petterle
}
