% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_ashw.R
\name{ASHW}
\alias{ASHW}
\alias{dashw}
\alias{pashw}
\alias{qashw}
\alias{rashw}
\title{Arcsecant Hyperbolic Weibull (ASHW) Distribution}
\usage{
dashw(x, mu, phi, tau = 0.5, log = FALSE)

pashw(q, mu, phi, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

qashw(p, mu, phi, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

rashw(n, mu, phi, tau = 0.5)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{quantile parameter (0 < mu < 1).}

\item{phi}{shape parameter (phi > 0).}

\item{tau}{the quantile level (0 < tau < 1). Defaults to 0.5 (median).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function, and random number generation for the 
Arcsecant Hyperbolic Weibull (ASHW) distribution parameterized by a quantile \code{mu} at level \code{tau}.
}
\examples{
# 1. Set parameters (Median parameterization)
mu_med <- 0.5
phi_val <- 1.5

# 2. Random generation
set.seed(123)
x <- rashw(1000, mu_med, phi_val, tau = 0.5)

# 3. Visualization
hist(x, prob = TRUE, main = "ASHW (Median 0.5)", col = "lightgray")
curve(dashw(x, mu_med, phi_val, tau = 0.5), add = TRUE, col = "red", lwd = 2)

# 4. Check P(X <= mu) = tau
p_mu <- pashw(mu_med, mu_med, phi_val, tau = 0.5)
cat("P(X <= mu) should be 0.5:", p_mu, "\n")

}
\references{
Kieschnick, R., & McCullough, B. D. (2003). Regression analysis of variates observed on (0, 1): percentages, proportions and fractions. 
\emph{Statistical Modelling}, \bold{3}(3), 193--213.
}
\author{
Ricardo Rasmussen Petterle
}
