% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_beta_mode.R
\name{BetaMode}
\alias{BetaMode}
\alias{dbeta_mode}
\alias{pbeta_mode}
\alias{qbeta_mode}
\alias{rbeta_mode}
\title{Beta Distribution (Mode Parameterization)}
\usage{
dbeta_mode(x, mu, phi, log = FALSE)

pbeta_mode(q, mu, phi, lower.tail = TRUE, log.p = FALSE)

qbeta_mode(p, mu, phi, lower.tail = TRUE, log.p = FALSE)

rbeta_mode(n, mu, phi)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{vector of mode parameters (0 < mu < 1).}

\item{phi}{vector of precision parameters (phi > 2).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function, and random number generation for the 
Beta distribution parameterized by the mode \code{mu} and precision \code{phi}.
}
\details{
The Beta distribution with mode \eqn{\mu} and precision \eqn{\phi} has shape parameters:
\deqn{\alpha = \mu(\phi - 2) + 1}
\deqn{\beta = (1 - \mu)(\phi - 2) + 1}
where \eqn{0 < \mu < 1} and \eqn{\phi > 2}.
}
\examples{
# 1. Set parameters
mu_val <- 0.7
phi_val <- 10

# 2. Generate random values
set.seed(123)
x <- rbeta_mode(1000, mu_val, phi_val)

# 3. Visualization
hist(x, prob = TRUE, main = "Beta (Mode 0.7)", col = "lightgray")
curve(dbeta_mode(x, mu_val, phi_val), add = TRUE, col = "red", lwd = 2)
abline(v = mu_val, col = "blue", lty = 2) # Mode line

}
\author{
Ricardo Rasmussen Petterle
}
