% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{BIC.unitregTMB}
\alias{BIC.unitregTMB}
\title{Extract Bayesian Information Criterion}
\usage{
\method{BIC}{unitregTMB}(object, ...)
}
\arguments{
\item{object}{An object of class \code{unitregTMB}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A numeric value representing the BIC.
}
\description{
Extracts the Bayesian Information Criterion (BIC) of a fitted \code{unitregTMB} model.
}
\examples{
\donttest{
# BIC(fit)
}
}
