% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_johnsonsb.R
\name{JohnsonSB}
\alias{JohnsonSB}
\alias{djohnsonsb}
\alias{pjohnsonsb}
\alias{qjohnsonsb}
\alias{rjohnsonsb}
\title{The Johnson SB Distribution}
\usage{
djohnsonsb(x, mu, phi, tau = 0.5, log = FALSE)

pjohnsonsb(q, mu, phi, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

qjohnsonsb(p, mu, phi, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

rjohnsonsb(n, mu, phi, tau = 0.5)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{quantile parameter (0 < mu < 1).}

\item{phi}{shape/precision parameter (phi > 0).}

\item{tau}{the quantile level (0 < tau < 1). Defaults to 0.5 (median).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function, and random number generation for the 
Johnson SB distribution parameterized by a quantile \code{mu} at level \code{tau}.
}
\examples{
# 1. Set parameters
mu_val <- 0.7
phi_val <- 1.0
tau_val <- 0.5

# 2. Generate random values
set.seed(123)
x <- rjohnsonsb(1000, mu_val, phi_val, tau_val)

# 3. Visualization
hist(x, prob = TRUE, main = "Johnson SB", col = "lightgray")
curve(djohnsonsb(x, mu_val, phi_val, tau_val), add = TRUE, col = "red", lwd = 2)

}
\references{
Johnson, N. L. (1949). Systems of frequency curves generated by methods of translation. 
\emph{Biometrika}, 36(1/2), 149-176.
}
\author{
Ricardo Rasmussen Petterle
}
