% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_kumaraswamy.R
\name{Kumaraswamy}
\alias{Kumaraswamy}
\alias{dkum_mode}
\alias{pkum_mode}
\alias{qkum_mode}
\alias{rkum_mode}
\alias{dkum_quantile}
\alias{pkum_quantile}
\alias{qkum_quantile}
\alias{rkum_quantile}
\title{Kumaraswamy Distribution}
\usage{
dkum_mode(x, mu, phi, log = FALSE)

pkum_mode(q, mu, phi, lower.tail = TRUE, log.p = FALSE)

qkum_mode(p, mu, phi, lower.tail = TRUE, log.p = FALSE)

rkum_mode(n, mu, phi)

dkum_quantile(x, mu, phi, tau = 0.5, log = FALSE)

pkum_quantile(q, mu, phi, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

qkum_quantile(p, mu, phi, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

rkum_quantile(n, mu, phi, tau = 0.5)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{location parameter (Mode or Quantile), depending on the function (0 < mu < 1).}

\item{phi}{shape parameter 'a' (phi > 0, note that phi > 1 for Mode parametrization).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{tau}{the quantile level to be modeled (only for quantile parametrization).}
}
\description{
Density, distribution function, quantile function, and random number generation for the 
Kumaraswamy distribution parameterized by the Mode or a Quantile.
}
\details{
The Kumaraswamy distribution has PDF:
\deqn{f(x) = a b x^{a-1} (1 - x^a)^{b-1}}
}
\references{
Kumaraswamy, P. (1980). A generalized probability density function for double-bounded random processes. 
\emph{Journal of Hydrology}, \bold{46}(1-2), 79--88.
}
\author{
Ricardo Rasmussen Petterle
}
