% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.unitregTMB}
\alias{logLik.unitregTMB}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{unitregTMB}(object, ...)
}
\arguments{
\item{object}{An object of class \code{unitregTMB}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class \code{logLik} containing the log-likelihood value, 
  along with attributes for degrees of freedom (\code{df}) and number of observations (\code{nobs}).
}
\description{
Extracts the log-likelihood of a fitted \code{unitregTMB} model.
}
