% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coef.R
\name{plot_coef}
\alias{plot_coef}
\title{Plot Coefficients from one or more unitregTMB models}
\usage{
plot_coef(
  ...,
  conf_level = 0.95,
  intercept = TRUE,
  component = "mu",
  ncol = 2,
  labels = NULL,
  parse_labels = FALSE,
  math_labels = FALSE,
  facet = TRUE,
  palette = "default",
  point_size = 2,
  strip_text_size = 11,
  axis_title_size = 11,
  axis_text_x_size = 10,
  axis_text_y_size = 12,
  vline_color = "gray40",
  point_color = "#0072B2",
  errorbar_color = "gray40",
  title = NULL,
  subtitle,
  xlab = "Estimate Value",
  ylab = "Coefficient"
)
}
\description{
Plot Coefficients from one or more unitregTMB models
}
