% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ranef.R
\name{plot_ranef}
\alias{plot_ranef}
\title{Plot Conditional Modes (Caterpillar Plot) for unitregTMB}
\usage{
plot_ranef(object, level = 0.95, ...)
}
\arguments{
\item{object}{A fitted \code{unitregTMB} model.}

\item{level}{Confidence level for the error bars. Default is 0.95.}

\item{...}{Additional graphical parameters passed to \code{plot}.}
}
\description{
Plot Conditional Modes (Caterpillar Plot) for unitregTMB
}
