% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ranef.R
\name{plot_ranef.unitregTMB}
\alias{plot_ranef.unitregTMB}
\title{Plot Conditional Modes (Caterpillar Plot) for unitregTMB}
\usage{
\method{plot_ranef}{unitregTMB}(object, level = 0.95, ...)
}
\arguments{
\item{object}{A fitted \code{unitregTMB} model.}

\item{level}{Confidence level for the error bars. Default is 0.95.}

\item{...}{Additional graphical parameters passed to \code{plot}.}
}
\value{
Invisibly returns a \code{data.frame} containing the sorted estimated random effects 
  (BLUPs), their standard errors, and the lower and upper bounds of the confidence intervals.
}
\description{
Produces a caterpillar plot of the conditional modes (BLUPs) of the random effects 
from a fitted \code{unitregTMB} model, along with their confidence intervals.
}
\examples{
\donttest{
# fit <- unitregTMB(y ~ age + bmi + gender + ipaq + (1 | id), 
#                   phi.formula = ~ 1,
#                   family = unitgamma(model_for = "mean"),  
#                   data = bodyfat_long)

# # Generate the caterpillar plot
# plot_ranef(fit, level = 0.95)
# 
# # You can also save the data used for the plot
# re_data <- plot_ranef(fit)
# head(re_data)
}

}
