% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.unitregTMB}
\alias{print.unitregTMB}
\title{Print a unitregTMB Model}
\usage{
\method{print}{unitregTMB}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{unitregTMB}.}

\item{digits}{The number of significant digits to use when printing.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the original \code{unitregTMB} object.
}
\description{
Prints the basic information and estimated coefficients of a fitted \code{unitregTMB} model.
}
