% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tex_summary.R
\name{print_tex_summary}
\alias{print_tex_summary}
\title{Print Regression Coefficients in LaTeX for unitregTMB}
\usage{
print_tex_summary(model, digits = 3, p_digits = 3, custom_labels = NULL)
}
\description{
Print Regression Coefficients in LaTeX for unitregTMB
}
