% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_simplex.R
\name{Simplex}
\alias{Simplex}
\alias{dsimplex}
\alias{psimplex}
\alias{qsimplex}
\alias{rsimplex}
\title{The Simplex Distribution}
\usage{
dsimplex(x, mu, phi, log = FALSE)

psimplex(q, mu, phi, lower.tail = TRUE, log.p = FALSE)

qsimplex(p, mu, phi, lower.tail = TRUE, log.p = FALSE)

rsimplex(n, mu, phi)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{mean parameter (0 < mu < 1).}

\item{phi}{dispersion parameter (phi > 0).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function, and random number generation for the 
Simplex distribution parameterized by the mean \code{mu} and dispersion \code{phi}.
}
\details{
The Simplex distribution is parameterized such that \eqn{\mu} is the mean 
and \eqn{\phi = \sigma^2} is the dispersion parameter. The variance of the 
distribution is approximately \eqn{Var(Y) \approx \phi \mu^3 (1-\mu)^3}.
}
\examples{
# 1. Set parameters
mu_val <- 0.5
phi_val <- 1.5

# 2. Generate random values
set.seed(123)
x <- rsimplex(1000, mu_val, phi_val)

# 3. Visualization
hist(x, prob = TRUE, main = "Simplex (Mean 0.5)", col = "lightgreen")
curve(dsimplex(x, mu_val, phi_val), add = TRUE, col = "darkgreen", lwd = 2)

}
\references{
Barndorff-Nielsen, O. E., & Jørgensen, B. (1991). Some parametric models on the simplex. 
\emph{Journal of Multivariate Analysis}, \bold{39}(1), 106--116.
}
\author{
Ricardo Rasmussen Petterle
}
