% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_ubs.R
\name{UBS}
\alias{UBS}
\alias{dubs}
\alias{pubs}
\alias{qubs}
\alias{rubs}
\title{Unit-Birnbaum-Saunders (UBS) Distribution}
\usage{
dubs(x, mu, phi, tau = 0.5, log = FALSE)

pubs(q, mu, phi, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

qubs(p, mu, phi, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

rubs(n, mu, phi, tau = 0.5)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{quantile parameter (0 < mu < 1).}

\item{phi}{shape parameter (phi > 0).}

\item{tau}{the quantile level (0 < tau < 1). Defaults to 0.5 (median).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function, and random number generation for the 
Unit-Birnbaum-Saunders (UBS) distribution parameterized by a quantile \code{mu} at level \code{tau}.
}
\examples{
# 1. Set parameters (Median)
mu_med <- 0.5
phi_val <- 0.5

# 2. Generate random values
set.seed(123)
x <- rubs(1000, mu_med, phi_val, tau = 0.5)

# 3. Visualization
hist(x, prob = TRUE, main = "UBS (Median 0.5)", col = "lavender")
curve(dubs(x, mu_med, phi_val, tau = 0.5), add = TRUE, col = "purple", lwd = 2)

}
\references{
Mazucheli, J., Menezes, A. F. B., & Dey, S. (2018). Unit-Birnbaum-Saunders distribution with applications. 
\emph{Chilean Journal of Statistics}, \bold{9}(1), 47--57.
}
\author{
Ricardo Rasmussen Petterle
}
