% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_unitgamma.R
\name{UnitGamma}
\alias{UnitGamma}
\alias{dugamma_mean}
\alias{pugamma_mean}
\alias{qugamma_mean}
\alias{rugamma_mean}
\alias{dugamma_mode}
\alias{pugamma_mode}
\alias{qugamma_mode}
\alias{rugamma_mode}
\title{Unit-gamma Distribution}
\usage{
dugamma_mean(x, mu, phi, log = FALSE)

pugamma_mean(q, mu, phi, lower.tail = TRUE, log.p = FALSE)

qugamma_mean(p, mu, phi, lower.tail = TRUE, log.p = FALSE)

rugamma_mean(n, mu, phi)

dugamma_mode(x, mu, phi, log = FALSE)

pugamma_mode(q, mu, phi, lower.tail = TRUE, log.p = FALSE)

qugamma_mode(p, mu, phi, lower.tail = TRUE, log.p = FALSE)

rugamma_mode(n, mu, phi)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{mean or mode parameter, depending on the function used (0 < mu < 1).}

\item{phi}{shape parameter (phi > 0).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function, and random number generation for the 
Unit-gamma distribution parameterized by the Mean or Mode.
}
\details{
The Unit-gamma distribution is obtained by the transformation \eqn{Y = \exp(-X)}, where \eqn{X \sim \Gamma(\text{shape}=\phi, \text{rate}=\beta)}.
}
\examples{
# 1. Mean Parameterization
mu_m <- 0.5
phi_m <- 2.0
x1 <- rugamma_mean(1000, mu_m, phi_m)
hist(x1, prob = TRUE, main = "Unit-gamma (Mean)", col = "lightgray", border = "white")
curve(dugamma_mean(x, mu_m, phi_m), add = TRUE, col = "red", lwd = 2)

# 2. Mode Parameterization
mu_d <- 0.7
phi_d <- 5.0
x2 <- rugamma_mode(1000, mu_d, phi_d)
hist(x2, prob = TRUE, main = "Unit-gamma (Mode)", col = "lightblue", border = "white")
curve(dugamma_mode(x, mu_d, phi_d), add = TRUE, col = "blue", lwd = 2)

}
\references{
Grassia, A. (1977). On a family of distributions with argument between 0 and 1 obtained by transformation of the gamma and beta families. 
\emph{Australian Journal of Statistics}, \bold{19}(2), 108--114.
}
\author{
Ricardo Rasmussen Petterle
}
