% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.unitregTMB}
\alias{vcov.unitregTMB}
\title{Extract Variance-Covariance Matrix}
\usage{
\method{vcov}{unitregTMB}(object, ...)
}
\arguments{
\item{object}{An object of class \code{unitregTMB}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A numeric matrix representing the estimated covariances between the parameter estimates.
}
\description{
Extracts the variance-covariance matrix of the main parameters of a fitted \code{unitregTMB} model.
}
